First commit of PINModel
Summary:
The goal of this project is to generate model classes for Objective-C (iOS) initially and
then expand to Java (Android) in the future.

The model definitions are stored in the json-schema v4 format but could potentially change to alternate intermediate representation types in the future (avro, proto, custom schema definition).

http://spacetelescope.github.io/understanding-json-schema/index.html

Goals of the model:
- Immutability : Model classes will be generated with Immutability as a
    requirement. Each class will have "Builder" classes that will aid in mutation
    by a copy-on-write policy.
- Type Safety: Based on the type information specified in the schema
    definition, each class will provide type validation and null reference checks
    to ensure model integrity.
- Custom Validation: Each property can specify a set of parameters that will be
    used for validation. Examples of these properties are defined in the
    json-schema v4 specification.