Integer enumeration support
Summary:
Adds support for "enum" keyword. The values of the enum array are json-schemas themselves and will include a "default" and "description" key.
Default indicates the integer value of this enum and description is the user-readable name that distinguishes a specific enum entry.

Example enum property:
"in_stock" : {
    "type": "integer",
    "enum": [
        { "default" : -1, "description" : "unknown" },
        { "default" : 0, "description" : "out_of_stock" },
        { "default" : 1, "description" : "in_stock" }
    ]
}

Test Plan: Generated code and it compiled

Reviewers: wendy, garrett

Differential Revision: https://phabricator.pinadmin.com/D56074