Use relative schema urls internally
As discussed in person, given that compilers should (if they have a
choice) produce bit-for-bit identical output, and a codegen tool like
plank is used in builds, it makes sense for `--print_deps` to output
dependencies via their relative paths.

Even if all things were equal and , this is still a nice change because:

1. The print_deps output is more readable and fewer bytes since absolute
paths are mostly repetitive and noisy
2. The logic for `decodeRef` can be simplified

I verified that Pinterest still builds correctly with this plank and
that we are still able to load schemas located in a different directory
(I tested with `../../foo.json`).