Initial support for Java (#106)
- This generation is targeted at Java running on the Android platform (Java 7).
It will be designed with that use-case in mind and may involve usage of
platform APIs that are not available in all Java environments
- Immutability is provided by AutoValue. This annotation is a dependency
of all models.
- Models assume you are using GSON and will generate SerializedName
annotations for model properties. In addition, we generate TypeAdapter methods to support auto-value-gson. This will likely be configurable to support Moshi in the future
- This commit does not have a working implementation of ADTs. This will
be addressed in a future commit and should not be relied upon