[Objective-C] Option for adding abilities to tell if a method is set.
Summary:
This is helpful for times where we have partially materialized model
objects. An instance of this might be that your API allows for per-field
request logic (like GraphQL) so you aren't guaranteed to have the value
of a model property unless it is requested.  This change aims to provide
a way for consumers of plank generated models to differentiate between
`nil` or default uninitialized values and ones that are actually
populated.

This means that we have...
- Three possible states for nullable properties (nonnull, null or unknown / not fetched)
- Two states for nonnull properties (nonnull or unknown / not fetched)
- Two states for primitive properties (some value or unknown / not fetched)

Tasks:
- Adds an isSet method for each property that utilizes the internal dirty properties bitmask to return the correct value.
= Updated integration tests to verify isSet is returning the correct
value