Emit BOOL properties as bitfields
This will make Pinterest models smaller in the heap, which is useful since they live for the life of the application. This changes only the model, not the short lived builders, to use bitfields. It requires special handling of boolean properites throughout the process.

In Pinterest, PIPin has 25 BOOLs (currently stored as 200 bytes, which takes 4 bytes afterwards), & PIUser has 40 (currently stored as 320 bytes, and takes 8 bytes afterwards). There may be even more wins, simply by grouping all the BOOLs near each other. If the layout of a class is a pointer, follwed by a BOOL, followed by a pointer, the 64 bit alignment of fields will cause that BOOL to actually be 8 bytes!