Only allocate _bits when necessary (#238)
We were previously allocating the _bits array as part of its
declaration. In many cases, we wouldn't use this memory; we would
just reassign _bits in the non-default constructors and builders.

Instead, we now only allocate _bits in the default Builder constructor.
We also now generate a default public model constructor when package-
private setters are enabled so that there's some backing storage to
support those set methods. (Note that we don't currently appear to
actually update _bits in those methods, which may be a separate issue.)

Lastly, we no longer need to read the "_bits" array in TypeAdapter#read,
so remove that case entirely.