Java equals() should evaluate primitives first
This is an optimization for generated equals() methods that evaluates
the cheaper primitive comparisons before the more expensive complex
object comparisons.

This change also renames Schema.isObjCPrimitiveType to the more general
Schema.isPrimitiveType. It was already being used by both Objective-C
and JavaScript, and it's logic can also be applied to Java.