Wrap -debugDescription in `#ifdef DEBUG`
This is a useful debugging method, but it doesn't need to be included as
part of a release build. If we discover a use case for that, we should
consider an appropriate -description implementation instead.

Because of the number of strings this method can use, especially for
models with many fields, this change results in a significant binary
size reduction for non-debug builds.