Add Objective-C type-checking support (#273)
We previously trusted the model dictionary to contain schema-compatible
types, but because that dictionary often comes from API-provided JSON,
it's prudent to take a "trust but verify" approach to initialization.

This change introduces an `initWithModelDictionary:error:` variant that
performs schema-based type checks before assigning a dictionary value to
a model property.

If the type check fails, the property is left unset (as if the value
never appeared in the model dictionary), and the provided `error`
argument is set to an `NSError` object. In `DEBUG` builds, this error
object will also contain a debug description that includes the
problematic key, expected type, and actual type of the last type failure
that occurred.

If the caller doesn't provided an `error` object, type checking is
skipped, and the model behaves as did previously. This is useful for
times when the model dictionary is trusted and the (minimal) performance
overhead introduced by the type checks is undesirable. This trade-off is
left to the caller.